bugfix: not all positions were selectable due to internal rounding errors
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@2258 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ plugins/goto/GotoDialog.cpp
@@ -20,7 +20,7 @@
 #include "GotoDialog.h"
 
 //***************************************************************************
-GotoDialog::GotoDialog(QWidget *widget, Mode mode, double pos,
+GotoDialog::GotoDialog(QWidget *widget, Mode mode, unsigned int pos,
                        double sample_rate, unsigned int signal_length)
     :QDialog(widget), Ui::GotoDlg()
 {
